+++ lib/src/extension.dart
@@ -1,0 +1,5 @@
+import 'stats.dart';
+
+extension StatsExtensions<T> on Iterable<num> {
+  Stats get stats => Stats.fromData(this);
+}