+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.0
 
+- Added `stats` extension property to `Iterable<num>`.
 - Added root mean square – `rms`.
+- Require Dart SDK `>=2.7.0 <3.0.0`.
 
 ## 0.2.0+3